Prefer `RUNNER_TEMP` for temporary directory (#50)
The directory which the `RUNNER_TEMP` env var points to is cleaned after
each job run and it's also the place where the runner/other actions put
their temporary files. Additionally, for sandboxed runners, the
temporary directoy (as returned by `tmpdir`) is likely not accessible by
the host's Nix, so the garbage collector thinks the profiles in this
directory are no longer alive and garbage-collects them.